+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/Container.js
@@ -28,7 +28,7 @@ import resolveEditableValue from '../../utils/editable-value/resolveEditableValu
 import {getCommonStyleByName} from '../../utils/getCommonStyleByName';
 import {getEditableLinkValue} from '../../utils/getEditableLinkValue';
 import {getFrontendTokenValue} from '../../utils/getFrontendTokenValue';
-import getLayoutDataItemClassName from '../../utils/getLayoutDataItemClassName';
+import getLayoutDataItemUniqueClassName from '../../utils/getLayoutDataItemUniqueClassName';
 import {getResponsiveConfig} from '../../utils/getResponsiveConfig';
 import {isValidSpacingOption} from '../../utils/isValidSpacingOption';
 import useBackgroundImageValue from '../../utils/useBackgroundImageValue';
@@ -169,7 +169,7 @@ const Container = React.forwardRef(
 			<HTMLTag
 				{...(link ? {} : data)}
 				className={classNames(className, {
-					[getLayoutDataItemClassName(
+					[getLayoutDataItemUniqueClassName(
 						item.itemId
 					)]: config.featureFlagLps132571,
 					[align]: !!align,
