+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/utils/getLayoutDataItemUniqueClassName.js
@@ -12,6 +12,6 @@
  * details.
  */
 
-export default function getLayoutDataItemClassName(itemId) {
+export default function getLayoutDataItemUniqueClassName(itemId) {
 	return `lfr-layout-structure-item-${itemId}`;
 }
